requirements.txt
alembic==1.8.1
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
Flask==2.2.2
Flask-CKEditor==0.4.6
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.30
protobuf==3.20.1
psycopg2==2.9.3
pycparser==2.21
PyMySQL==1.0.2
SQLAlchemy==1.4.41
Werkzeug==2.2.2
WTForms==3.0.1
